only updating labels that are correctly labeled

+++ adversaries.py
@@ -105,7 +105,7 @@ class LBFGS(object):
       # For each label in outputs that is correctly classified, replace
       # use second best class. Otherwise, stick with current prediction
       new_labels = Variable(predicted_classes).masked_scatter_(predicting_correct_class,
-                                                     second_best_class.float())
+                                                               second_best_class)
 
     else:
       new_labels = Variable(torch.LongTensor([target_class]*self.batch_size))
